backend/requirements.txt
# Django Core
Django==4.2.16
djangorestframework==3.14.0
django-filter==23.5
django-environ==0.11.2
django-cors-headers==4.9.0

# Database
PyMySQL==1.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.2

# Crawling & Data Processing
Scrapy==2.11.2
pandas==2.2.2
numpy==1.26.4

# Task Queue
celery[redis]==5.4.0
redis==5.0.8
django-celery-beat==2.6.0

# Media Processing
Pillow==10.4.0

# Text Processing
bleach==6.1.0

# WSGI Server
gunicorn==22.0.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.32.3

# Development Tools
django-extensions==3.2.3
